add support for multiple metadata versions
Version 10 Support
Version 9 Support
