+++ core/src/metadata.rs
@@ -19,6 +19,8 @@
 
 /// Different metadata versions
 mod versions;
+mod version_10;
+mod version_09;
 
 use codec::{Decode, Encode, EncodeAsRef, HasCompact};
 use failure::Fail;
@@ -26,9 +28,9 @@ use runtime_metadata_latest::{
     DecodeDifferent, RuntimeMetadata, RuntimeMetadataPrefixed, StorageEntryModifier,
     StorageEntryType, StorageHasher, META_RESERVED,
 };
+
 use std::{
     collections::{HashMap, HashSet},
-    convert::TryFrom,
     marker::PhantomData,
     rc::Rc,
     str::FromStr,
@@ -372,111 +374,14 @@ pub enum Error {
     ExpectedDecoded,
     #[fail(display = "Invalid Event {}:{}", _0, _1)]
     InvalidEventArg(String, &'static str),
+    #[fail(display = "Deserialize: {}", _0)]
+    Serde(#[fail(cause)] serde_json::Error)
 }
 
-impl TryFrom<RuntimeMetadataPrefixed> for Metadata {
-    type Error = Error;
-
-    fn try_from(metadata: RuntimeMetadataPrefixed) -> Result<Self, Self::Error> {
-        if metadata.0 != META_RESERVED {
-            Err(Error::InvalidPrefix)?;
-        }
-        let meta = match metadata.1 {
-            RuntimeMetadata::V10(meta) => meta,
-            _ => Err(Error::InvalidVersion)?,
-        };
-        let mut modules = HashMap::new();
-        let mut modules_by_event_index = HashMap::new();
-        let mut event_index = 0;
-        for (i, module) in convert(meta.modules)?.into_iter().enumerate() {
-            let module_name = convert(module.name.clone())?;
-            let module_metadata = convert_module(i, module)?;
-            // modules with no events have no corresponding definition in the
-            // top level enum
-            if !module_metadata.events.is_empty() {
-                modules_by_event_index.insert(event_index, module_name.clone());
-                event_index = event_index + 1;
-            }
-            modules.insert(module_name, Rc::new(module_metadata));
-        }
-        Ok(Metadata {
-            modules,
-            modules_by_event_index,
-        })
+impl From<serde_json::Error> for Error {
+    fn from(err: serde_json::Error) -> Error {
+        Error::Serde(err)
     }
-}
-
-fn convert<B: 'static, O: 'static>(dd: DecodeDifferent<B, O>) -> Result<O, Error> {
-    match dd {
-        DecodeDifferent::Decoded(value) => Ok(value),
-        _ => Err(Error::ExpectedDecoded),
-    }
-}
-
-fn convert_module(
-    index: usize, module: runtime_metadata_latest::ModuleMetadata,
-) -> Result<ModuleMetadata, Error> {
-    let mut storage_map = HashMap::new();
-    if let Some(storage) = module.storage {
-        let storage = convert(storage)?;
-        let prefix = convert(storage.prefix)?;
-        for entry in convert(storage.entries)?.into_iter() {
-            let entry_name = convert(entry.name.clone())?;
-            let entry_prefix = format!("{} {}", prefix, entry_name);
-            let entry = convert_entry(entry_prefix, entry)?;
-            storage_map.insert(entry_name, entry);
-        }
-    }
-    let mut call_map = HashMap::new();
-    if let Some(calls) = module.calls {
-        for (index, call) in convert(calls)?.into_iter().enumerate() {
-            // HERE modify
-            let name = convert(call.name)?;
-            call_map.insert(name, vec![index as u8]);
-        }
-    }
-    let mut event_map = HashMap::new();
-    if let Some(events) = module.event {
-        for (index, event) in convert(events)?.into_iter().enumerate() {
-            event_map.insert(index as u8, convert_event(event)?);
-        }
-    }
-    Ok(ModuleMetadata {
-        index: index as u8,
-        name: convert(module.name)?,
-        storage: storage_map,
-        calls: call_map,
-        events: event_map,
-    })
-}
-
-fn convert_event(
-    event: runtime_metadata_latest::EventMetadata,
-) -> Result<ModuleEventMetadata, Error> {
-    let name = convert(event.name)?;
-    let mut arguments = HashSet::new();
-    for arg in convert(event.arguments)? {
-        let arg = arg.parse::<EventArg>()?;
-        arguments.insert(arg);
-    }
-    Ok(ModuleEventMetadata { name, arguments })
-}
-
-fn convert_entry(
-    prefix: String, entry: runtime_metadata_latest::StorageEntryMetadata,
-) -> Result<StorageMetadata, Error> {
-    let default = convert(entry.default)?;
-    let documentation = convert(entry.documentation)?;
-    Ok(StorageMetadata {
-        prefix,
-        modifier: entry.modifier,
-        ty: entry.ty,
-        default,
-        documentation: documentation
-            .iter()
-            .map(|s| s.to_string())
-            .collect::<Vec<String>>(),
-    })
 }
 
 #[cfg(test)]
